backend/requirements.txt
pyjwt==2.8.0
flask==3.0.0
pytest==8.0.0
flask_sqlalchemy==3.0.3
werkzeug==3.0.1
cryptography==41.0.7
